Ensure we use the correct bucket / prefix
We use the submodul'ed template for EKS, so that we can control its
version. However, if we don't pass through `QSS3BucketName`,
`QSS3BucketRegion` & `QSS3KeyPrefix`, the nested templates would default
to the upstreams bucket (`aws-quickstart`).
Depending on the submodul'ed version in this repo & the version in the
upstream bucket, this can lead to all kinds of incompatibility.