Install aws-cfn-bootstrap as early as reasonably possible
Moving things around a bit, so that we are able to install the
aws-cfn-bootstrap tools a bit and thus set up the trap a bit earlier.

Signaling the error or success of the cloud-init run now becomes quite
important, otherwise the `UbuntuBastion` will just block until
`CreationPolicy`'s timeout expires.