Workaround for source-controller issues
NOTE: this is just a workaround and should be removed as soon as
possible

The source-controller that ships with TAP 1.5.x (version 0.7.0) has a
bug with auth against ECR via IAM roles.
Therefore we need to roll back to a previous version. The version from
TAP 1.4.3 (version 0.6.3) seems to work.

So we:
- add the TAP 1.4.3 repo
   - relocate it, if that what the user has configured
- add an overlay to tap which pins the source-controller package to
  0.6.3